requirements.txt
dbt-duckdb>=1.8
duckdb>=1.1
streamlit>=1.40
requests>=2.31
pyyaml>=6.0
apscheduler>=3.10
pandas>=2.1
plotly>=5.0
python-dotenv>=1.0
